+++ application/Components/Auth.php
@@ -10,6 +10,7 @@ namespace App\Components;
 
 use App\Entity;
 use App\Libraries\Constant;
+use App\Repository\User\UserStatus;
 
 use Rid\Base\Component;
 use Rid\Helpers\JWTHelper;
@@ -71,7 +72,7 @@ class Auth extends Component
         if ($user_id !== false && is_int($user_id) && $user_id > 0) {
             $user_id = intval($user_id);
             $curuser = app()->site->getUser($user_id);
-            if ($curuser->getStatus() !== Entity\User::STATUS_DISABLED)  // user status shouldn't be disabled
+            if ($curuser->getStatus() !== UserStatus::DISABLED)  // user status shouldn't be disabled
                 return $curuser;
         }