+++ application/Repository/Torrent/TorrentType.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: Rhilip
+ * Date: 1/11/2020
+ * Time: 2020
+ */
+
+declare(strict_types=1);
+
+namespace App\Repository\Torrent;
+
+
+class TorrentType
+{
+    const SINGLE = 'single';
+    const MULTI = 'multi';
+}